pyproject.toml
[project]
name = "testenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.4",
    "strands-agents>=1.4.0",
    "strands-agents-builder>=0.1.7",
    "strands-agents-tools>=0.2.3",
]
